--- .gitignore
@@ -1,2 +1,0 @@
-## 폴더 무시
-design/
