+++ .idea/modules.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/server/src/main/main.iml" filepath="$PROJECT_DIR$/server/src/main/main.iml" />
+      <module fileurl="file://$PROJECT_DIR$/server/src/test/test.iml" filepath="$PROJECT_DIR$/server/src/test/test.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/wedding-team-management.iml" filepath="$PROJECT_DIR$/.idea/wedding-team-management.iml" />
+    </modules>
+  </component>
+</project>
