--- server/ex.txt
@@ -1,1 +1,0 @@
-폴더 구조 올리기용 지워도 됨
