+++ server/management/src/main/java/lavi/management/ManagementApplication.java
@@ -1,0 +1,13 @@
+package lavi.management;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ManagementApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ManagementApplication.class, args);
+	}
+
+}
