+++ server/management/src/main/java/lavi/management/domain/Gender.java
@@ -1,0 +1,5 @@
+package lavi.management.domain;
+
+public enum Gender {
+    MAN, WOMAN
+}
