+++ server/management/src/main/java/lavi/management/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package lavi.management.repository;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class MemberRepository {
+}
