+++ server/management/src/main/java/lavi/management/service/MemberService.java
@@ -1,0 +1,7 @@
+package lavi.management.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class MemberService {
+}
