+++ server/management/src/main/resources/application.properties
@@ -1,0 +1,8 @@
+spring.datasource.url=jdbc:mysql://localhost:3306/lavi
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.username=root
+spring.datasource.password=Su702020!
+
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.hibernate.ddl-auto=create
