+++ server/management/src/test/java/lavi/management/ManagementApplicationTests.java
@@ -1,0 +1,13 @@
+package lavi.management;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ManagementApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
