--- server/src/main/java/lavi/attendance/AttendanceApplication.java
@@ -1,13 +1,0 @@
-package lavi.attendance;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AttendanceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AttendanceApplication.class, args);
-	}
-
-}
