--- server/src/main/java/lavi/attendance/domain/member/MemoryMemberRepository.java
@@ -1,41 +1,0 @@
-package lavi.attendance.domain.member;
-
-import org.springframework.stereotype.Repository;
-
-import java.util.*;
-
-@Repository
-public class MemoryMemberRepository implements MemberRepository {
-
-    private static final Map<Long, Member> store = new HashMap<>();
-    private static long sequence = 0L;
-
-    @Override
-    public Member save(Member member) {
-        member.setId(sequence++);
-        store.put(member.getId(), member);
-        return member;
-    }
-
-    @Override
-    public Member findById(Long id) {
-        return store.get(id);
-    }
-
-    @Override
-    public Optional<Member> findByLoginId(String loginId) {
-        return findAll().stream()
-                .filter(member -> member.getLoginId().equals(loginId))
-                .findFirst();
-    }
-
-    @Override
-    public List<Member> findAll() {
-        return new ArrayList<>(store.values());
-    }
-
-    @Override
-    public void clearStore() {
-        store.clear();
-    }
-}
