--- server/src/test/java/lavi/attendance/AttendanceApplicationTests.java
@@ -1,13 +1,0 @@
-package lavi.attendance;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AttendanceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
